+++ core/src/cell/builder.rs
@@ -9,7 +9,8 @@ use num_traits::{One, Zero};
 use crate::cell::dict::{DictBuilder, ValWriter};
 use crate::cell::error::{MapTonCellError, TonCellError};
 use crate::cell::{ArcCell, Cell, CellParser};
-use crate::TonAddress;
+use crate::tlb_types::traits::TLBObject;
+use crate::{TonAddress, TonHash};
 
 pub(crate) const MAX_CELL_BITS: usize = 1023;
 pub(crate) const MAX_CELL_REFERENCES: usize = 4;
@@ -189,8 +190,8 @@ impl CellBuilder {
 
     /// Stores address without optimizing hole address
     pub fn store_raw_address(&mut self, val: &TonAddress) -> Result<&mut Self, TonCellError> {
-        self.store_u8(2, 0b10u8)?;
-        self.store_bit(false)?;
+        self.store_u8(2, 0b10u8)?; //store as MsgAddressInt
+        self.store_bit(false)?; // always no anycast
         let wc = (val.workchain & 0xff) as u8;
         self.store_u8(8, wc)?;
         self.store_slice(val.hash_part.as_slice())?;
@@ -290,6 +291,24 @@ impl CellBuilder {
         Ok(self)
     }
 
+    pub fn store_tlb<T: TLBObject>(&mut self, obj: &T) -> Result<&mut Self, TonCellError> {
+        obj.write(self)?;
+        Ok(self)
+    }
+
+    pub fn store_tlb_optional<T: TLBObject>(
+        &mut self,
+        maybe_obj: Option<&T>,
+    ) -> Result<&mut Self, TonCellError> {
+        if let Some(obj) = maybe_obj {
+            self.store_bit(true)?;
+            self.store_tlb(obj)?;
+        } else {
+            self.store_bit(false)?;
+        }
+        Ok(self)
+    }
+
     // https://docs.ton.org/develop/data-formats/tl-b-types#maybe
     pub fn store_maybe_cell_ref(
         &mut self,
@@ -340,6 +359,10 @@ impl CellBuilder {
             );
             self.store_reference(&dict_data)
         }
+    }
+
+    pub fn store_tonhash(&mut self, ton_hash: &TonHash) -> Result<&mut Self, TonCellError> {
+        self.store_slice(ton_hash.as_slice())
     }
 
     pub fn remaining_bits(&self) -> usize {
@@ -425,6 +448,7 @@ mod tests {
     use crate::cell::dict::predefined_readers::{key_reader_u8, val_reader_uint};
     use crate::cell::{CellBuilder, TonCellError};
     use crate::types::TonAddress;
+    use crate::TonHash;
 
     #[test]
     fn test_extend_and_invert_bits() -> Result<(), TonCellError> {
@@ -707,6 +731,21 @@ mod tests {
         let mut parser = cell.parser();
         let parsed = parser.load_dict(8, key_reader_u8, val_reader_uint)?;
         assert_eq!(data, parsed);
+        Ok(())
+    }
+
+    #[test]
+    fn test_store_tonhash() -> Result<(), TonCellError> {
+        let mut writer = CellBuilder::new();
+        let ton_hash =
+            TonHash::from_hex("9f31f4f413a3accb706c88962ac69d59103b013a0addcfaeed5dd73c18fa98a8")?;
+
+        writer.store_tonhash(&ton_hash)?;
+        let cell = writer.build()?;
+        let mut parser = cell.parser();
+        let parsed = parser.load_tonhash()?;
+        assert_eq!(ton_hash, parsed);
+        parser.ensure_empty()?;
         Ok(())
     }
 }